renamed graph component
