+++ BeaconAndLoaves/ClientApp/src/components/pages/OwnerDashboard/OwnerDashboard.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import DatePicker from 'react-datepicker';
-import Graph2 from '../../Graph/Graph2';
+import Graph from '../../Graph/Graph';
 import rentalRequests from '../../../helpers/data/rentalRequests';
 import userRequests from '../../../helpers/data/userRequests';
 import './OwnerDashboard.scss';
@@ -54,18 +54,6 @@ class OwnerDashboard extends React.Component {
         this.figureTotal();
       });
   }
-
-  // getPropertiesWithRentalTotals = (properties) => {
-  //   const { allRentals } = this.state;
-  //   properties.forEach((property) => {
-  //     rentalRequests.getAllRentalsByPropertyIdWithTotals(property.id)
-  //       .then((rentals) => {
-  //         rentals.forEach((rental) => {
-  //           allRentals.push(rental);
-  //         });
-  //       });
-  //   });
-  // }
 
   getSinglePropertyCreatedOn = () => {
     const { selectedProperty, properties } = this.state;
@@ -205,7 +193,7 @@ class OwnerDashboard extends React.Component {
             <p>Average ${averagePerRental} per rental</p>
           </div>
       </div>
-        <Graph2
+        <Graph
           currentUser={this.props.currentUser}
           toggleGraphModal={this.toggleGraphModal}
           graphModal={graphModal}
